+++ CHANGELOG.md
@@ -2,6 +2,35 @@
 
 All notable changes to `statamic-responsive-images` will be documented in this file
 
+## v3.0.0 - 2023-02-04
+
+**This is a breaking change release, as such please see UPGRADE.md for instructions on how to upgrade from 2.x version to 3.x.**
+
+### What's Changed
+
+- Extensible dimension calculators by @ncla in https://github.com/spatie/statamic-responsive-images/pull/193
+  - Developers now can customize the amount of images that get generated and their exact dimensions through a custom `DimensionCalculator` that developers bind in `ServiceProvider`. Calculations can be done for each breakpoint and source, and receive `Breakpoint` / `Source` in the calculation methods, which allows to access breakpoint parameters, original asset and more. For motivation and examples of this, please see the PR and the original issue associated with it.
+- Fix max widths config value or glide width param not being respected in some cases
+- Fix floating numbers being output for width and height values for Glide endpoint, they are now rounded integers
+- Placeholder now can be toggled per breakpoint
+- `<img>` tag src will now go through Glide instead of pointing to the original asset file, due to `DimensionCalculators` ability to return specific widths and heights for `<img>` tag
+- Empty `media` attributes will no longer be outputted
+- Properties have been renamed to be more descriptive:
+  - `value` to `minWidth`
+  - `unit` to `widthUnit`
+- Add `mimeType` property to explicitly help browsers determine what images does the `<source>` contain
+- Housekeeping: add additional tests for dimension calculator and GraphQL
+
+**Full Changelog**: https://github.com/spatie/statamic-responsive-images/compare/v2.15.2...v3.0.0
+
+## v2.15.2 - 2023-01-28
+
+### What's Changed
+
+- Fix broken thumbnails on control panel entry index pages on Statamic v3.4 by @ncla in https://github.com/spatie/statamic-responsive-images/pull/201
+
+**Full Changelog**: https://github.com/spatie/statamic-responsive-images/compare/v2.15.1...v2.15.2
+
 ## v2.15.1 - 2023-01-12
 
 ### What's Changed